Use fixture for test_events
